fix(#3): persist sidebar state across refresh and reset on login
